fish: optimize config to improve startup time :)

+++ fish/config.fish
@@ -1,60 +1,75 @@
 # CAREFUL: Make sure you're editing this file '.dots', not in its usual folder
 
 # path setup
-fish_add_path -p /usr/local/sbin /usr/local/bin /usr/bin ~/.cargo/bin ~/.deno/bin /usr/local/go/bin
+if not contains -- /usr/local/sbin $PATH
+    fish_add_path -p /usr/local/sbin /usr/local/bin /usr/bin ~/.cargo/bin ~/.deno/bin /usr/local/go/bin
+end
 
 # {{#if dotter.packages.mac}}
-eval (/opt/homebrew/bin/brew shellenv)
+if test -f /opt/homebrew/bin/brew
+    eval (/opt/homebrew/bin/brew shellenv &)
+end
 # {{/if}}
 
 # set env variable for XDG_CONFIG_HOME
-set -gx XDG_CONFIG_HOME ~/.config/
+if not set -q XDG_CONFIG_HOME
+    set -gx XDG_CONFIG_HOME ~/.config/
+end
 
 # set editor and visual env variables
-set -gx EDITOR {{editor}};
-set -gx VISUAL {{editor}};
+if not set -q EDITOR
+    set -gx EDITOR {{editor}}
+    set -gx VISUAL {{editor}}
+end
 
-# app abbreviations
-abbr -a -- hx   {{editor}};
-abbr -a -- G    'gitui'
-abbr -a -- rm   '"Nope, use `rip` instead."'
+if status is-interactive
+    # app abbreviations
+    abbr -a -- hx   {{editor}}
+    abbr -a -- G    'gitui'
+    abbr -a -- rm   '"Nope, use `rip` instead."'
 
-# enable starting zellij with a keyboard shortcut.
-# loads a previous session if one is available,
-# otherwise just starts a new session
-# (only works outside of zellij, fortunately)
-bind \cz 'zellij a || zellij; commandline -f repaint'
+    # enable starting zellij with a keyboard shortcut
+    bind \cz 'zellij a || zellij; commandline -f repaint'
 
-# bindings for navigating word-by-word
-bind \eh backward-word
-bind \el forward-word
+    # bindings for navigating word-by-word
+    bind \eh backward-word
+    bind \el forward-word
+end
 
-# put openai and anthropic API keys into env
-set -gx OPENAI_API_TOKEN (cat ~/.openai)
-set -gx ANTHROPIC_API_TOKEN (cat ~/.anthropic)
+# put API keys into env 
+load_token OPENAI_API_TOKEN ~/.openai &
+load_token ANTHROPIC_API_TOKEN ~/.anthropic &
 
 # atuin
-atuin init fish --disable-up-arrow | source
+if not set -q ATUIN_SESSION; and type -q atuin
+    atuin init fish --disable-up-arrow | source &
+end
 
 # {{#if dotter.packages.work}}
 # ----------------------------
 #  WORK STUFF
 # ----------------------------
 # initiate fnm (used to manage node installs)
-fnm env | source
+if not set -q FNM_DIR; and type -q fnm
+    fnm env | source &
+end
 
-# transfer gh and jira tokens from file to env
-set -gx GITHUB_TOKEN (cat ~/.github_token)
-set -gx JIRA_API_TOKEN (cat ~/.jira_token)
-set -gx COPILOT_API_KEY (cat ~/.copilot)
+# put API keys into env 
+load_token GITHUB_TOKEN ~/.github_token &
+load_token JIRA_API_TOKEN ~/.jira_token &
+load_token COPILOT_API_KEY ~/.copilot &
 
-# work kubernetes-related abbrevs
-abbr -a -- kdev  'k9s --context=aws-dev'
-abbr -a -- kqa   'k9s --context=aws-qa'
-abbr -a -- kprod 'k9s --context=aws-prod'
+if status is-interactive
+    # work kubernetes-related abbrevs
+    abbr -a -- k9s\
+        kdev='k9s --context=aws-dev'\
+        kqa='k9s --context=aws-qa'\
+        kprod='k9s --context=aws-prod'
+end
 # {{/if}}
 
 # {{#if dotter.packages.linux}}
-source ~/.xprofile
+if test -f ~/.xprofile
+    source ~/.xprofile
+end
 # {{/if}}
-#
